pyproject.toml
[tool.poetry]
name = "test-case-generation"
version = "0.1.0"
description = ""
authors = ["Ayah Alsaad <147486757+ayahalsaad@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "0.27.1"
uvloop = "0.19.0"
watchfiles = "0.21.0"
websocket-client = "1.7.0"
websockets = "12.0"
wrapt = "1.16.0"
zipp = "3.17.0"
annotated-types = "0.6.0"
anyio = "4.3.0"
asgiref = "3.7.2"
backoff = "2.2.1"
bcrypt = "4.1.2"
build = "1.0.3"
cachetools = "5.3.2"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
chroma-hnswlib = "0.7.3"
chromadb = "0.4.23"
click = "8.1.7"
coloredlogs = "15.0.1"
deprecated = "1.2.14"
distro = "1.9.0"
exceptiongroup = "1.2.0"
fastapi = "0.109.2"
filelock = "3.13.1"
flatbuffers = "23.5.26"
fsspec = "2024.2.0"
google-auth = "2.28.1"
googleapis-common-protos = "1.62.0"
grpcio = "1.62.0"
h11 = "0.14.0"
httpcore = "1.0.4"
httptools = "0.6.1"
httpx = "0.27.0"
huggingface-hub = "0.20.3"
humanfriendly = "10.0"
idna = "3.6"
importlib-metadata = "6.11.0"
importlib-resources = "6.1.1"
kubernetes = "29.0.0"
mmh3 = "4.1.0"
monotonic = "1.6"
mpmath = "1.3.0"
numpy = "1.26.4"
oauthlib = "3.2.2"
onnxruntime = "1.17.0"
openai = "1.12.0"
opentelemetry-api = "1.23.0"
opentelemetry-exporter-otlp-proto-common = "1.23.0"
opentelemetry-exporter-otlp-proto-grpc = "1.23.0"
opentelemetry-instrumentation = "0.44b0"
opentelemetry-instrumentation-asgi = "0.44b0"
opentelemetry-instrumentation-fastapi = "0.44b0"
opentelemetry-proto = "1.23.0"
opentelemetry-sdk = "1.23.0"
opentelemetry-semantic-conventions = "0.44b0"
opentelemetry-util-http = "0.44b0"
orjson = "3.9.15"
overrides = "7.7.0"
packaging = "23.2"
posthog = "3.4.2"
protobuf = "4.25.3"
pulsar-client = "3.4.0"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pydantic = "2.6.2"
pydantic-core = "2.16.3"
pypika = "0.48.9"
pyproject-hooks = "1.0.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.36.3"
sympy = "1.12"
tenacity = "8.2.3"
tokenizers = "0.15.2"
tomli = "2.0.1"
tqdm = "4.66.2"
typer = "0.9.0"
typing-extensions = "4.9.0"
urllib3 = "2.2.1"
langchain = "^0.1.9"
pypdf2 = "^3.0.1"
langchain-openai = "^0.0.7"
pdfminer-six = "^20231228"
pymongo = "^4.6.2"
argparse = "^1.4.0"
lxml = "^5.1.0"
bs4 = "^0.0.2"
tiktoken = "^0.6.0"
pypdf = "^4.0.2"
python-multipart = "^0.0.9"
chardet = "^5.2.0"
faiss-cpu = "^1.8.0"
humps = "^0.2.2"
pyhumps = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


